+++ testing/devtest/cap-screen.py
@@ -1,6 +1,0 @@
-# (c) Copyright 2020 by Coinkite Inc. This file is covered by license found in COPYING-CC.
-#
-import sim_display
-
-RV.write(sim_display.full_contents)
-